refactor: add `ArtworkModalHeaderInfo` component
