+++ src/Apps/CollectorProfile/Routes/Saves2/Components/CreateNewListModal/CreateNewListModal.tsx
@@ -2,10 +2,7 @@ import { Formik, FormikHelpers } from "formik"
 import createLogger from "Utils/logger"
 import { ModalDialog } from "@artsy/palette"
 import { useTranslation } from "react-i18next"
-import {
-  ArtworkEntity,
-  CreateNewListModalHeader,
-} from "./CreateNewListModalHeader"
+import { CreateNewListModalHeader } from "./CreateNewListModalHeader"
 import { useCreateCollection } from "Apps/CollectorProfile/Routes/Saves2/Components/Actions/Mutations/useCreateCollection"
 import { FC } from "react"
 import { useTracking } from "react-tracking"
@@ -16,6 +13,7 @@ import {
   ArtworkListFormikValues,
   validationSchema,
 } from "Apps/CollectorProfile/Routes/Saves2/Components/ArtworkListForm/ArtworkListForm"
+import { ArtworkEntity } from "Components/Artwork/ManageArtworkForSaves"
 
 export interface ArtworkList {
   internalID: string
