+++ src/Apps/CollectorProfile/Routes/Saves2/Components/CreateNewListModal/CreateNewListModalHeader.tsx
@@ -1,11 +1,7 @@
-import { Flex, Spacer, Text } from "@artsy/palette"
-import { SavesEntityImage } from "Apps/CollectorProfile/Routes/Saves2/Components/SavesEntityImage"
+import { Spacer } from "@artsy/palette"
+import { ArtworkModalHeaderInfo } from "Apps/CollectorProfile/Routes/Saves2/Components/ArtworkModalHeaderInfo"
+import { ArtworkEntity } from "Components/Artwork/ManageArtworkForSaves"
 import { FC } from "react"
-
-export interface ArtworkEntity {
-  title: string
-  imageURL: string | null
-}
 
 interface CreateNewListModalHeaderProps {
   artwork: ArtworkEntity
@@ -15,10 +11,9 @@ export const CreateNewListModalHeader: FC<CreateNewListModalHeaderProps> = ({
   artwork,
 }) => {
   return (
-    <Flex flex={1} flexDirection="row" alignItems="center" mb={2}>
-      <SavesEntityImage size={50} url={artwork.imageURL} />
-      <Spacer x={1} />
-      <Text lineClamp={2}>{artwork.title}</Text>
-    </Flex>
+    <>
+      <ArtworkModalHeaderInfo artwork={artwork} />
+      <Spacer y={2} />
+    </>
   )
 }
