+++ src/Apps/CollectorProfile/Routes/Saves2/Components/SelectArtworkListsModal/SelectArtworkListsHeader.tsx
@@ -1,5 +1,5 @@
-import { Button, Flex, Message, Spacer, Text } from "@artsy/palette"
-import { SavesEntityImage } from "Apps/CollectorProfile/Routes/Saves2/Components/SavesEntityImage"
+import { Button, Flex, Message, Spacer } from "@artsy/palette"
+import { ArtworkModalHeaderInfo } from "Apps/CollectorProfile/Routes/Saves2/Components/ArtworkModalHeaderInfo"
 import {
   ModalKey,
   useManageArtworkForSavesContext,
@@ -19,23 +19,13 @@ export const SelectArtworkListsHeader: FC = () => {
     })
   }
 
-  const titleElements = [
-    artwork.artists && `${artwork.artists}, `,
-    <i>{artwork.title}</i>,
-    artwork.year && `, ${artwork.year}`,
-  ]
-
   return (
     <>
       <Flex
         flexDirection={["column", "row"]}
         alignItems={["stretch", "center"]}
       >
-        <Flex flex={1} flexDirection="row" alignItems="center">
-          <SavesEntityImage size={50} url={artwork.imageURL} />
-          <Spacer x={1} />
-          <Text lineClamp={2}>‘{titleElements}’</Text>
-        </Flex>
+        <ArtworkModalHeaderInfo artwork={artwork} />
 
         <Spacer x={[0, 1]} y={[2, 0]} />
